删除goflow中的chrome渲染，修正依赖问题
