--- gocodefuncs/screenshot_test.go
@@ -1,140 +1,0 @@
-package gocodefuncs
-
-import (
-	"context"
-	"github.com/LubyRuffy/goflow/utils"
-	"github.com/chromedp/chromedp"
-	"github.com/stretchr/testify/assert"
-	"log"
-	"os"
-	"testing"
-	"time"
-)
-
-func Test_chromeActions(t *testing.T) {
-	buf := []byte{}
-	type args struct {
-		in      chromeActionsInput
-		logf    func(string, ...interface{})
-		timeout int
-		actions []chromedp.Action
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "测试正常screen picture",
-			args: args{
-				in: chromeActionsInput{
-					URL: "https://www.baidu.com",
-				},
-				logf:    func(s string, i ...interface{}) {},
-				timeout: 10,
-				actions: []chromedp.Action{
-					chromedp.Sleep(time.Second * time.Duration(1)),
-					chromedp.CaptureScreenshot(&buf),
-				},
-			},
-		},
-		{
-			name: "测试自定义proxy & UA",
-			args: args{
-				in: chromeActionsInput{
-					URL:       "http://www.baidu.com",
-					Proxy:     "socks5://127.0.0.1:7890",
-					UserAgent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Edg/91.0.864.59",
-				},
-				logf:    func(s string, i ...interface{}) {},
-				timeout: 10,
-				actions: []chromedp.Action{
-					chromedp.Sleep(time.Second * time.Duration(1)),
-					chromedp.CaptureScreenshot(&buf),
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := chromeActions(tt.args.in, func(s string, i ...interface{}) {}, tt.args.timeout, tt.args.actions...)
-			assert.Nil(t, err)
-			assert.NotNil(t, buf)
-			t.Logf("screenshot result: %s", buf[0:5])
-		})
-	}
-}
-
-func TestAddObjectSlice(t *testing.T) {
-	runner := newTestRunner(t, "")
-	type args struct {
-		p          Runner
-		objectName string
-		ele        string
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "测试添加object环境",
-			args: args{
-				p:          runner,
-				objectName: "test",
-				ele:        "test1",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			p := tt.args.p
-			AddObjectSlice(p, tt.args.objectName, tt.args.ele)
-			object, ok := p.GetObject(tt.args.objectName)
-			assert.True(t, ok)
-			assert.Equal(t, tt.args.ele, object.([]string)[0])
-		})
-	}
-}
-
-func TestAddUrlToTitle(t *testing.T) {
-	type args struct {
-		url          string
-		hasTimeStamp bool
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		{
-			name: "测试截图添加url地址",
-			args: args{url: `https://fofa.info`, hasTimeStamp: false},
-		},
-		{
-			name: "测试截图添加url地址 & 时间戳",
-			args: args{url: `https://fofa.info`, hasTimeStamp: true},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx, cancel := chromedp.NewContext(
-				context.Background(),
-			)
-			defer cancel()
-
-			var buf []byte
-			err := chromedp.Run(ctx, fullScreenshot(tt.args.url, 90, &buf))
-			assert.Nil(t, err)
-
-			gotResult, err := AddUrlToTitle(tt.args.url, buf, tt.args.hasTimeStamp)
-			assert.Nil(t, err)
-			assert.Greater(t, len(gotResult), len(buf))
-
-			// 效果展示
-			var fn string
-			fn, err = utils.WriteTempFile(".png", func(f *os.File) error {
-				_, err = f.Write(gotResult)
-				return err
-			})
-			log.Printf("save modified pic into: %s", fn)
-		})
-	}
-}
